Parsing refactoring
* Parsing modules are moved to Parsing directory in containers
* Class Parser is deleted because it is useless
* Parser's functions refactored into 3 arrow functions, written in DynamicLoading.js, StaticLoading.js and Loading.js, to simplify data exchange between this methods and Header
